Fix optimistic locking issue
